+++ RELEASES.md
@@ -27,13 +27,13 @@
 * Subscription shipping preferences
 
 == V2.7.0
-* Send Order.ip address to authorize.net
+* Send BazaarCore::Order.ip address to authorize.net
 * Attach failed transactions to users
 * Set email and first name on cart_item post
 * Quantity or equals one on cart_item post
 
 == V2.7.0
-* Send Order.ip address to authorize.net
+* Send BazaarCore::Order.ip address to authorize.net
 * Attach failed transactions to users
 * Set email and first name on cart_item post
 * Quantity or equals one on cart_item post